qwertz

+++ core/src/uas/lntv/pacmangame/Screens/SplashScreen.java
@@ -76,7 +76,7 @@ public class SplashScreen implements Screen {
         this.renderer = new OrthogonalTiledMapRenderer(TMX_MAP);
         this.CAM = new OrthographicCamera();
         this.gamePort = new FitViewport(MAP_WIDTH * TILE_SIZE, MAP_HEIGHT * TILE_SIZE, CAM);
-        this.CAM.position.set((float)(MAP_WIDTH * TILE_SIZE) / 2, (float)(MAP_WIDTH * TILE_SIZE) / 2, 0);
+        this.CAM.position.set((MAP_WIDTH * TILE_SIZE) / 2f, (MAP_WIDTH * TILE_SIZE) / 2f, 0);
 
         this.layerLNTV = (TiledMapTileLayer) TMX_MAP.getLayers().get("Walls");
         this.layerGDX = (TiledMapTileLayer) TMX_MAP.getLayers().get("Path");
